Modified project structure to use the atomic design system
