+++ client/src/components/atoms/avatar/avatar.tsx
@@ -1,0 +1,14 @@
+type Props = {
+  url: string;
+  text: string;
+};
+const Avatar = ({ url, text }: Props) => {
+  return (
+    <figure>
+      <img src={url} alt={text} />
+      <span className='sr-only'>{text}</span>
+    </figure>
+  );
+};
+
+export { Avatar };
