+++ client/src/components/atoms/avatar/styles.tsx
@@ -1,0 +1,1 @@
+export {};
