+++ client/src/components/atoms/index.ts
@@ -1,2 +1,3 @@
 export * from './buttons';
 export * from './logo';
+export * from './styles';
