+++ client/src/components/atoms/logo.tsx
@@ -1,14 +1,12 @@
 import { links } from '@src/common';
 import { Link } from 'react-router-dom';
+import { Avatar } from './avatar/avatar';
 
 const Logo = () => {
   return (
-    <figure>
-      <Link to={'/'}>
-        <img src={links?.logo} alt='Audiophile Logo' />
-        <span className='sr-only'>Audiophile Logo</span>
-      </Link>
-    </figure>
+    <Link to={'/'}>
+      <Avatar url={links?.logo} text={'Audiophile Logo'} />
+    </Link>
   );
 };
