+++ client/src/components/pages/earphones.tsx
