+++ client/src/components/pages/index.ts
