+++ client/src/components/pages/styles.tsx
