--- client/src/pages/headphones.tsx
@@ -1,7 +1,0 @@
-type Props = {};
-
-const Headphones = (props: Props) => {
-  return <h1>Headphones</h1>;
-};
-
-export { Headphones };
